% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_aws_creds}
\alias{get_aws_creds}
\title{Resolve AWS credentials}
\usage{
get_aws_creds()
}
\description{
Wrapper around aws.signature::locate_credentials() with simple error messaging.
}
\keyword{internal}
