% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{resolve_bedrock_endpoint}
\alias{resolve_bedrock_endpoint}
\title{Resolve Bedrock Agent Runtime host + region}
\usage{
resolve_bedrock_endpoint(region = NULL)
}
\description{
Resolve Bedrock Agent Runtime host + region
}
\keyword{internal}
