% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\arguments{
\item{x, y}{Values to coalesce.}
}
\value{
Either \code{x} (when present) or \code{y}.
}
\description{
Return \code{y} when \code{x} is \code{NULL} or of length 0; otherwise return \code{x}.
}
\details{
This is an internal helper used throughout the package.
}
\keyword{internal}
